split up database file
